+++ todo.md
@@ -1,0 +1,5 @@
+### 📝 TODO
+
+- [ ] 코스 신청 서버 액션 실제 구현 (현재 임시 구현)
+- [ ] 코스 신청 완료 후 상태 확인 페이지 구현
+- [ ] 코스 난이도 ENUM 값 데이터베이스와 동기화 필요
